Disable jshint warnings for constructors
The jshint rule "nonew" raises the warning "Do not use 'new'
for side effects" when a constructor is called and the created
instance is not used.

When creating Tabris widgets using constructors, there are
cases that don't require any further references to the created
widget. Remove this check from the jshint config.

Change-Id: I699b218c4111b1650a204697efe77e2d4cdafe12